requirements.txt
-i https://pypi.org/simple/
click==7.0
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
pillow==6.2.1
pytesseract==0.3.0
tesseract==0.1.3
werkzeug==0.16.0
